partitioning code in a dedicate file
