+++ src/lib.rs
@@ -2,3 +2,4 @@
 pub mod impg;
 pub mod seqidx;
 pub mod paf;
+pub mod partition;
